+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- acl - Principals add by user/group uuid possible 
+- acl - Principals dropdown searchable with groups/user
+
 ## [2.32.7] - 2022-09-29
 
 ### Changed